// frontend/src/app/api/model/elevationProfilePoint.ts
/**
 * Life Management 2.0 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Egy mintapont a magassági profilban (backlog/tura-utvonaltervezo/103-... 2.3 fázis).
 */
export interface ElevationProfilePoint { 
    /**
     * Táv az útvonal kezdőpontjától méterben.
     */
    distanceMeters: number;
    /**
     * Tengerszint feletti magasság méterben.
     */
    elevationMeters: number;
}